fix: settings formatting
Temporarily align the format of settings to include the account/product
in the namespaces. This is to align to the way the current occurrence
construction code works, and also to the way composite settings are
generated.

Also defualt the tenant is only one tenant is found and no tenant is
provided.