fix: default link processing
Current link processing treats the target as present even if it is in
a different region. Currently the cmdb input processing limits the
stack files it returns to the current account and region, but this
results in the link process failing (as state lookups fail).

Once placements are correctly handled, and the input filter is switched
when generating the occurrence for a link target, the current cmdb
input processing will be correct.

But for now, return all stacks regardless of region for a segment.